Add support for hooks
Further customization is available after downloading a tool version
using hooks. Hooks are Groovy scripts that can be run after downloading
the tool version to customize it. An example is to set a custom
`cacerts` file after a Java download, or set a custom `settings.xml`
file for Maven.

This is done by:
1. Create a `hooks` folder in the same path as the jshim.jar file
2. Inside the `hooks` folder, write a `<tool>-postdownload.groovy` file

The following variables are injected into the Groovy script:
- `TOOL_NAME`: the tool name
- `TOOL_VERSION`: the tool version
- `TOOL_HOME`: the home folder for the tool version

An example of a postdownload groovy script can be found in:
`examples/hooks/maven-postdownload.groovy`